+++ src/recode.jl
@@ -1,7 +1,5 @@
 const ≅ = isequal
 
-import Missings: replace
-
 """
     recode!(dest::AbstractArray, src::AbstractArray[, default::Any], pairs::Pair...)
 
@@ -369,10 +367,4 @@ function recode(a::CategoricalArray{S, N, R}, default::Any, pairs::Pair...) wher
         dest = CategoricalArray{Missings.T(T), N, R}(size(a))
     end
     recode!(dest, a, default, pairs...)
-end
-
-function Missings.replace(a::CategoricalArray{S, N, R, V, C}, replacement::V) where {S, N, R, V, C}
-    pool = CategoricalPool(a.pool)
-    v = C(get!(pool, replacement), pool)
-    Missings.replace(a, v)
 end
