+++ test/16_recode.jl
@@ -534,12 +534,4 @@ end
     @test !isordered(y)
 end
 
-@testset "Missings.replace should work on CategoricalArrays" begin
-    x = categorical(["a", "b", missing, "a"])
-    y = categorical(["a", "b", "", "a"])
-    r = Missings.replace(x, "")
-    @test isa(r, Missings.EachReplaceMissing)
-    @test y == CategoricalArray(collect(r))
-end
-
 end
